refactor: path alias 적용
