+++ biome.json
@@ -47,6 +47,7 @@
       },
       "nursery": {
         "useSortedClasses": {
+          "level": "warn",
           "options": {
             "attributes": ["classList"],
             "functions": ["clsx", "cva", "tw", "tw.*"]
