+++ src/features/shared/icons/icons.tsx
@@ -1,5 +1,5 @@
+import { IconMap, type IconMapTypes, type IconSizeTypes, IconSizes } from "@/features/shared/icons";
 import type React from "react";
-import { IconMap, type IconMapTypes, type IconSizeTypes, IconSizes } from "../icons";
 
 interface SVGIconProps extends React.SVGProps<SVGSVGElement> {
   title?: string;
