+++ src/features/shared/icons/images/file.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from "../";
+import type { IconProps } from "@/features/shared/icons";
 
 const FileEditIcon = ({
   title,
