+++ src/features/survey/create/components/question/dropDownQuestion.tsx
@@ -1,11 +1,11 @@
 import SVGIcon from "@/features/shared/icons/icons";
+import Options from "@/features/survey/create/components/options";
 import { useSurveyStore } from "@/features/survey/create/store/survey";
 import { deleteOption } from "@/lib/utils/createPageUtils";
 import { handleOptionDragEnd } from "@/lib/utils/handleDragEnd";
 import isModeDisabled from "@/lib/utils/isModeDisabled";
 import type { Option, Question, QuestionProps } from "@/types";
 import { DragDropContext, Draggable, Droppable } from "@hello-pangea/dnd";
-import Options from "../options";
 
 const DropDownQuestion: React.FC<QuestionProps> = ({ question, mode, onResponseChange }) => {
   const _isDisabled = isModeDisabled(mode);
