+++ src/features/survey/create/components/question/starRatingQuestion.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import SVGIcon from "@/features/shared/icons/icons";
+import Options from "@/features/survey/create/components/options";
 import { useSurveyStore } from "@/features/survey/create/store/survey";
 import type { QuestionProps } from "@/types";
 import { useEffect, useState } from "react";
-import Options from "../options";
 
 const StarRatingQuestion: React.FC<QuestionProps> = ({ question, mode }) => {
   const { updateQuestion } = useSurveyStore();
